Se agregaron mensajes de alerta al eliminar desde perfil

+++ public/assets/js/academico.js
@@ -55,7 +55,24 @@ $(document).ready(function(){
             }
         });
 
-        $.ajax({
+        if (!confirm("ADVERTENCIA!! va a proceder a eliminar este registro, si desea eliminarlo de click en ACEPTAR\n de lo contrario de click en CANCELAR.")) {
+            return false;
+            }
+            else {
+                $.ajax({
+                    type: "DELETE",
+                    url: 'deleteacad/' + idacad,
+                    success: function (data) {
+                        console.log(data);
+                        $("#academico" + idacad).remove();
+                    },
+                    error: function (data) {
+                        console.log('Error:', data);
+                    }
+                });
+            }
+
+        /*$.ajax({
             type: "DELETE",
             url: 'deleteacad/' + idacad,
             success: function (data) {
@@ -65,7 +82,7 @@ $(document).ready(function(){
             error: function (data) {
                 console.log('Error:', data);
             }
-        });
+        });*/
 
         $("#erroresContentC").html(errHTML); 
         $('#erroresModalC').modal('show');
